Time: 11 ms (6.92%), Space: 23.6 MB (22.95%) - LeetHub

+++ 0155-min-stack/0155-min-stack.cpp
@@ -1,7 +1,7 @@
 class MinStack {
 public:
     stack<long long>s;
-    int m;
+    long long m;
     MinStack() {
         
     }
